refactor: inline ripgrep into agentbinary
